# cmd/scanner/docs/swagger.yaml
basePath: /
definitions:
  api.ApiItem:
    properties:
      item:
        type: object
    type: object
  api.ApiItems:
    properties:
      items:
        type: object
    type: object
  api.ApiWithItem:
    properties:
      apiVersion:
        type: string
      data:
        type: object
    type: object
  api.OnlyAccountRes:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.OnlyFlagRes:
    properties:
      flag:
        type: boolean
    type: object
...
  model.GlobalRejectPolicy:
    properties:
      cicd:
        type: boolean
      k8s_deployment:
        type: boolean
      mode:
        type: string
      online_monitor:
        type: boolean
    type: object
...
  model.ImageWhitelist:
    properties:
      created_at:
        type: string
      digest:
        type: string
      full_repo_name:
        type: string
      id:
        type: integer
      library:
        type: string
      tag:
        type: string
    type: object
  model.KVHash:
    properties:
      en:
        type: KeyValue
      zh:
        type: KeyValue
    type: object
  model.KVHashs:
    properties:
      KVHash:
        $ref: '#/definitions/model.KVHash'
        type: object
    type: object
  model.NotifyContext:
    properties:
      Cluster:
        type: string
      CustomKV:
        items:
          $ref: '#/definitions/model.KVHashs'
        type: array
      Namespace:
        type: string
      PodName:
        type: string
      PodUID:
        type: string
      ServiceID:
        type: string
    type: object
...
  model.RejectOnlineMoniterImage:
    properties:
      digest:
        type: string
      image:
        type: string
      notify_context:
        $ref: '#/definitions/model.NotifyContext'
        type: object
      type:
        type: string
    type: object
...
  model.ReportImgBackInfo:
    properties:
      created:
        type: string
      created_by:
        type: string
      image_digest:
        type: string
      malicious:
        items:
          type: string
        type: array
      pkgs:
        items:
          type: string
        type: array
      sensitive_files:
        items:
          type: string
        type: array
      vulus:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: { }
  description: This is the Vegeta central server - Scanner
  license: { }
  title: Vegeta API·
  version: "1.0"
paths:
  /api/v1/imagereject/online_moniter:
    post:
      description: k8s&在线监控生成接口
      parameters:
        - description: JSON数据
          in: body
          name: body
          required: true
          schema:
            items:
              allOf:
                - $ref: '#/definitions/model.RejectOnlineMoniterImage'
                - properties:
                    NotifyContext:
                      allOf:
                        - $ref: '#/definitions/model.NotifyContext'
                        - properties:
                            CustomKV:
                              items:
                                allOf:
                                  - $ref: '#/definitions/model.KVHashs'
                                  - properties:
                                      KVHash:
                                        $ref: '#/definitions/model.KVHash'
                                    type: object
                              type: array
                          type: object
                  type: object
            type: array
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          item:
                            $ref: '#/definitions/api.OnlyFlagRes'
                        type: object
                type: object
      summary: TickOnlineScan
      tags:
        - image reject
...
  /api/v1/imagereject/policy/global:
    get:
      description: 获取全局策略
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          item:
                            $ref: '#/definitions/model.GlobalRejectPolicy'
                        type: object
                type: object
      summary: GetGlobalPolicy
      tags:
        - image reject
    put:
      description: 更新全局策略
      parameters:
        - description: 全局策略JSON数据
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/model.GlobalRejectPolicy'
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    $ref: '#/definitions/api.ApiItem'
                type: object
      summary: UpdateGlobalPolicy
      tags:
        - image reject
  /api/v1/imagereject/policy/single:
    get:
      description: 获取策略列表
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItems'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/model.RejectPolicy'
                            type: array
                        type: object
                type: object
      summary: ListPolicy
      tags:
        - image reject
    post:
      description: 新增策略接口
      parameters:
        - description: 策略JSON数据
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/model.RejectPolicy'
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    $ref: '#/definitions/api.ApiItem'
                type: object
      summary: CreatePolicy
      tags:
        - image reject
  /api/v1/imagereject/policy/single/:id:
    delete:
      description: 删除单条策略
      parameters:
        - description: policy ID
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    $ref: '#/definitions/api.ApiItem'
                type: object
      summary: DeletePolicy
      tags:
        - image reject
    put:
      description: 更新策略接口
      parameters:
        - description: JSON数据
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/model.RejectPolicy'
        - description: 策略ID
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    $ref: '#/definitions/api.ApiItem'
                type: object
      summary: UpdatePolicy
      tags:
        - image reject
...
  /api/v1/imagereject/whitelist:
    post:
      description: 创建白名单
      parameters:
        - description: JSON数据
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/model.ImageWhitelist'
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          item:
                            $ref: '#/definitions/model.ImageWhitelist'
                        type: object
                type: object
      summary: CreateWhitelist
      tags:
        - image reject
  /api/v1/imagereject/whitelist/:id:
    delete:
      description: 删除白名单
      parameters:
        - description: 白名单ID
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    $ref: '#/definitions/api.ApiItem'
                type: object
      summary: DeleteWhitelist
      tags:
        - image reject
  /api/v1/images/:imgDigest/layers:
    get:
      description: 获取镜像各层级的信息
      parameters:
        - description: imgDigest
          in: query
          name: imgDigest
          required: true
          type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/model.ReportImgBackInfo'
                            type: array
                        type: object
                type: object
      summary: ListImgLayers
      tags:
        - scan image
  /api/v1/images/app/:imageID/bases:
    get:
      description: 获取应用镜像的基础镜像列表
      parameters:
        - description: 镜像ID
          in: path
          name: imageID
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/model.ImageResponse'
                            type: array
                        type: object
                type: object
      summary: ListAppToBaseImage
      tags:
        - image reject
...
  /api/v1/images/base/:imageID/apps:
    get:
      description: 获取基础镜像的应用镜像列表
      parameters:
        - description: 镜像ID
          in: path
          name: imageID
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/model.ImageResponse'
                            type: array
                        type: object
                type: object
      summary: ListBaseToAppImage
      tags:
        - image reject
...
  /api/v1/register/registries:
    get:
      description: 获取registry列表信息
      parameters:
        - description: 是否需要配置策略的仓库
          in: query
          name: no_policy
          required: true
          type: boolean
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          item:
                            $ref: '#/definitions/api.OnlyAccountRes'
                        type: object
                type: object
      summary: 获取仓库列表
      tags:
        - registry
  /api/v1/register/registry:
    get:
      description: 获取registry具体信息
      parameters:
        - description: 仓库类型
          in: query
          name: usetype
          required: true
          type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          item:
                            $ref: '#/definitions/api.OnlyAccountRes'
                        type: object
                type: object
      summary: 获取指定仓库的具体信息
      tags:
        - registry
...
  /api/v1/scan/harbor/scanOneStatus:
    get:
      description: 获取单个镜像的扫描状态
      parameters:
        - description: Image ID
          in: query
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItem'
                      - properties:
                          item:
                            $ref: '#/definitions/model.ScanOneStatusResponse'
                        type: object
                type: object
      summary: GetScanOneStatus
      tags:
        - scan image
...
  /api/v1/vulns/all:
    get:
      description: 获取漏洞列表
      parameters:
        - description: 'for vuln like '
          in: query
          name: search
          type: string
        - description: int
          in: query
          name: offset
          required: true
          type: integer
        - description: int
          in: query
          name: limit
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/api.ApiWithItem'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/api.ApiItems'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/model.VulnList'
                            type: array
                        type: object
                type: object
      summary: all
      tags:
        - Vuln
...
swagger: "2.0"